requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bootstrap4==0.1.0
Django==3.1.4
django-bootstrap4==2.3.1
django-filter==2.4.0
django-forms-bootstrap==3.1.0
django-mathfilters==1.0.0
djangorestframework==3.12.2
importlib-metadata==2.1.1
Markdown==3.3.3
mod-wsgi==4.7.1
Pillow==8.0.1
pytz==2020.4
soupsieve==2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
zipp==3.4.0
